+++ cli/CHANGELOG.md
@@ -13,11 +13,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Account effects can be seen
   - All effects can be seen
   - All ledgers can be seen
+  - All operations can be seen
   - All payments can be seen
   - All trades can be seen
   - Ledger effects can be seen
   - Ledger transactions can be seen
   - Ledger payments can be seen
+  - Operation effects can be seen
   - Trade aggregations can be seen
   - Transaction details can be seen
   - Transaction effects can be seen